Cleanup configs
